Change packerbuilds to not use util script
packerbuilds-util script has been deleted. The main reasoning behind
this is because I usually just fill in the parameters when running
Jenkins jobs that are not from cron. Defaults are nice but this seems
to be harder to maintain as this project gets bigger.

Also, README.md has been updated to give notice that the
guide is currently out of date. Eventually I will update the README.