Add the functionality of shell preprocessing
From a comment in the code:

Packer supports the notion of local shell post processing and
provisioning (this occurs after the VM boots, regardless whether the
shell code runs on the VM or not) on the local machine but not before
processing the packer template. This is in attempt to replicate the
idea of shell pre processing (or preprocessing).