Add the notion of project and basic builds
This update to jenkins-packerbuilds now allows for the
distinction of basic/project builds. To make this happen,
the UI for jenkins-packerbuilds had to be changed a good bit.

Mainly, when a project build is selected to be done, I wanted
the UI to be smart enough to 'hide' some parameters, if those
parameters were mean't to stay a certain value given a
particular project. Several of the parameters are now
Active Choice Reactive Reference parameters to be able to
hide when needed or noted by the project.

Job parameter values have be migrated over to a single
cfg/conf file. This was done because hardcoding the values
into each parameter file was bad and proned to error over
time. Each parameter file now calls on the specfic value env
var as needed.

While hiding of some the parameters, now exists the notion
that a particular build may now require more than one
provisioner. The packerbuiolds script has now been updated
to be able to move over 'N' number of provisioners based
on what is specified in the packer template.